refactor(editor): renaming folder name from "inspector" to "tiles" so it actually reflect what it is.
