+++ Assets/Code/NonebNi/LevelEditor/Di/TileInspectorModule.cs
@@ -1,7 +1,7 @@
 ﻿using NonebNi.Core.Coordinates;
 using NonebNi.Core.Level;
-using NonebNi.LevelEditor.Level.Data;
-using NonebNi.LevelEditor.Level.Inspector;
+using NonebNi.LevelEditor.Level.Maps;
+using NonebNi.LevelEditor.Level.Tiles;
 using StrongInject;
 using UnityUtils.Factories;
