+++ Assets/Code/NonebNi/LevelEditor/Level/Entities/EditorEntityPositioningService.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using NonebNi.Core.Coordinates;
 using NonebNi.EditorComponent.Entities;
-using NonebNi.LevelEditor.Level.Data;
+using NonebNi.LevelEditor.Level.Maps;
 using UnityEngine;
 
 namespace NonebNi.LevelEditor.Level.Entities
