+++ Assets/Code/NonebNi/LevelEditor/Level/LevelEditorModel.cs
@@ -1,4 +1,4 @@
-﻿using NonebNi.LevelEditor.Level.Data;
+﻿using NonebNi.LevelEditor.Level.Maps;
 
 namespace NonebNi.LevelEditor.Level
 {
