+++ Assets/Code/NonebNi/LevelEditor/Level/Maps/EditorNode.cs
@@ -6,7 +6,7 @@ using NonebNi.Core.Tiles;
 using NonebNi.EditorComponent.Entities;
 using UnityEngine;
 
-namespace NonebNi.LevelEditor.Level.Data
+namespace NonebNi.LevelEditor.Level.Maps
 {
     [Serializable]
     public partial class EditorNode
