+++ Assets/Code/NonebNi/LevelEditor/Level/Maps/GridPresenter.cs
@@ -1,6 +1,4 @@
-﻿using NonebNi.LevelEditor.Level.Data;
-
-namespace NonebNi.LevelEditor.Level.Maps
+﻿namespace NonebNi.LevelEditor.Level.Maps
 {
     public class GridPresenter
     {
@@ -8,14 +6,14 @@ namespace NonebNi.LevelEditor.Level.Maps
         private readonly LevelEditorModel _levelEditorModel;
         private readonly NonebEditorModel _nonebEditorModel;
 
-        public bool IsDrawingGrid => _nonebEditorModel.IsGridVisible;
-        public IEditorMap Map => _levelEditorModel.EditorMap;
-
         public GridPresenter(GridView gridView, LevelEditorModel levelEditorModel, NonebEditorModel nonebEditorModel)
         {
             _gridView = gridView;
             _levelEditorModel = levelEditorModel;
             _nonebEditorModel = nonebEditorModel;
         }
+
+        public bool IsDrawingGrid => _nonebEditorModel.IsGridVisible;
+        public IEditorMap Map => _levelEditorModel.EditorMap;
     }
 }
