+++ Assets/Code/NonebNi/LevelEditor/Level/Settings/LevelSavingService.cs
@@ -1,6 +1,4 @@
-﻿using NonebNi.LevelEditor.Level.Data;
-
-namespace NonebNi.LevelEditor.Level.Settings
+﻿namespace NonebNi.LevelEditor.Level.Settings
 {
     public class LevelSavingService
     {
