+++ Assets/Code/NonebNi/LevelEditor/NonebEditor.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using NonebNi.LevelEditor.Common.Events;
 using NonebNi.LevelEditor.Di;
-using NonebNi.LevelEditor.Level.Data;
+using NonebNi.LevelEditor.Level;
 using NonebNi.LevelEditor.Toolbar;
 using StrongInject;
 using UnityEditor;
